fixed rendering boiiiiiiiiii

+++ src/Loader.h
@@ -10,13 +10,13 @@ class Loader {
 public:
     ~Loader();
 
-    GLuint createVAO(std::vector<double> &vertexes, std::vector<double>& normals, std::vector<double>& tex, std::vector<unsigned int> &indices);
+    GLuint createVAO(std::vector<float> &vertexes, std::vector<float>& normals, std::vector<float>& tex, std::vector<unsigned int> &indices);
 
 private:
     GLuint genVao();
     void unbindVAO();
 
-    void storeDataInAttributeList(GLuint index, int coordSize, const std::vector<double> &data);
+    void storeDataInAttributeList(GLuint index, int coordSize, const std::vector<float> &data);
 
     void bindIndicesBuffer(const std::vector<unsigned int> &indices);
 };
